edit: WritePage Radio Button 이동
